+++ spec/Suite/RegistrySpec.php
@@ -211,6 +211,18 @@ describe('Registry', function() {
             expect($di->offsetGet(Engine::class))->toBeAnInstanceOf(V8::class);
         });
 
+        it('calls factory with make function that creates an instance and resolve it\'s parameters from deleagte', function() {
+            $delegate = new Registry;
+            $delegate->offsetSet(Engine::class, V8::class);
+
+            $di = new Registry([], $delegate);
+            $di->offsetSet(Engine::class, function($c, $make) {
+                return $make(W16::class);
+            });
+
+            expect($di->offsetGet(Engine::class))->toBeAnInstanceOf(W16::class);
+        });
+
         it('calls factory with make function that when called without parameters it creates an instance of the abstract', function() {
             $di = new Registry;
             $di->offsetSet(V8::class, function($c, $make) {